Moved ICan to new file
